I changed all the ints to longs and now I can safely do somewhere around 10 digits without overflowing. Still need to take care of display overflow in a sensible way however.

+++ Calculator/CalculatorViewController.h
@@ -9,7 +9,7 @@
 #import <UIKit/UIKit.h>
 
 int operation;
-int selectedNumber;
+long selectedNumber;
 float runningTotal;
 
 @interface CalculatorViewController : UIViewController
